// src/api/generated/model/hourPreferenceData.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */
import type { HourPreferenceDataTimePeriod } from './hourPreferenceDataTimePeriod';
import type { HourPreferenceDataType } from './hourPreferenceDataType';

export interface HourPreferenceData {
  /**
   * The date of the preference
   * @nullable
   */
  date?: string | null;
  /**
   * The day of the week, 0 is Sunday, 6 is Saturday
   * @minimum 0
   * @maximum 6
   */
  dayOfWeek: number;
  /** The time period */
  timePeriod: HourPreferenceDataTimePeriod;
  /** The preference type, RECURRENT is a recurring preference, ONE_TIME is a one time preference */
  type: HourPreferenceDataType;
}