// src/api/generated/model/hourPreferenceDataTimePeriod.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Ludora API
 * API for the Ludora app
 * OpenAPI spec version: 0.0.1
 */

/**
 * The time period
 */
export type HourPreferenceDataTimePeriod =
  (typeof HourPreferenceDataTimePeriod)[keyof typeof HourPreferenceDataTimePeriod];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const HourPreferenceDataTimePeriod = {
  MORNING: 'MORNING',
  AFTERNOON: 'AFTERNOON',
  EVENING: 'EVENING',
} as const;